chore: moved location for dialog
