+++ app/routes/settings/account/index.tsx
@@ -18,7 +18,7 @@ import { Button } from "~/components/ui/button";
 import { Form, FormControl, FormField, FormItem, FormMessage } from "~/components/ui/form";
 import { Input } from "~/components/ui/input";
 
-import WarningComponent from "~/components/warning-dialog";
+import WarningComponent from "~/components/ui/warning-dialog";
 
 import { ChevronRight, TriangleAlert, type LucideIcon } from "lucide-react";
